examples/requirements.txt
# assume torch is already installed

# pycolmap for data parsing
git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e

imageio[ffmpeg]
numpy<2.0.0
scikit-learn
tqdm
torchmetrics[image]
opencv-python-headless
tyro>=0.8.8
Pillow
tensorboard
tensorly
pyyaml
matplotlib
fsspec==2024.10.0
torchvision==0.16.0
